Add an input library
A GameWidget doesn't have an event filter anymore. It is now up to the
developer to provide input handlers to entities.

The new Keyboard component provides key events handling.

With this addition comes a FrameAction component, which is basically a
function called right after input processing. For example, it is useful
to create a camera controller.